web/agents/willy_website_manager/requirements.txt
google-adk
python-dotenv
google-generativeai
google-cloud-aiplatform
